requirements.txt
cat > requirements.txt << 'EOF'
flask
yfinance
pandas
python-dotenv
requests
EOF